client/requirements.txt
requests==2.22.0
Pillow==6.1.0